!271 变更SqlmqQueue执行逻辑变更为按照执行序列号顺序执行

+++ src/main/java/cn/cerc/db/redis/RedisData.java
@@ -32,7 +32,7 @@ public class RedisData {
         return items();
     }
 
-    public RedisData setExpire(int seconds) {
+    public RedisData setExpire(long seconds) {
         timestamp = System.currentTimeMillis() + seconds * 1000;
         return this;
     }
